refactor(search): Cardの一元化とDialogの分離
